# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Product:
    description: Product defintes the structure for an API product
    properties:
      description:
        description: the description of the product
        type: string
        x-go-name: Description
      id:
        description: the id for this user
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: the name for this user
        type: string
        x-go-name: Name
      price:
        description: the price of the product
        format: float
        type: number
        x-go-name: Price
      sku:
        description: Product SKU (Stock Keeping Unit)
        example: '"abc-def-ghi"'
        type: string
        x-go-name: SKU
    required:
    - name
    - price
    - sku
    type: object
    x-go-package: github.com/rasyad91/introMicroservices/data
info:
  description: Documentation for Product API
  title: of Product API
  version: 1.0.0
paths:
...
  /products/{id}:
    delete:
      description: Deletes a product from the database
      operationId: deleteProduct
      parameters:
      - description: The id of the product to delete from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
      tags:
      - products
    put:
      description: Updates a product from the database
      operationId: updateProduct
      responses:
        "200":
          $ref: '#/responses/productsResponseWrapper'
      tags:
      - products
produces:
- application/json
responses:
  noContent:
    description: ""
  productsResponseWrapper:
    description: ""
    schema:
      $ref: '#/definitions/Product'
schemes:
- http
swagger: "2.0"